All links and redirects are now relative
WARNING - this might break. We'll revert this commit if so.
The rationale is for those using proxies where the host name is not proxied
on, as in Apache's ProxyPreserveHost.
RFC2161 does not permit relative Location: header, but an upcoming revision
stil in-draft will apparently allow relative spec. I believe all browsers to
support relative Location: headers already.