[#181] AP Client Count report was broken
The DB query was joining on the device_port_wireless table which as any fule
kno isn't using the device port subinterface name so there would always be
zero results (as wireless data is related to the subinterface).

Fixed the query to use the device_port_ssid table, and also added a dedicated
template so additional data can be displayed in the report.