update NodeVendor.pm
It would seem like we'd want for this to be distinct mac addresses here right? Otherwise you'd could get duplicate counters. In our environment where things move around this makes for repeat lines in the nodes table. To get a true inventory count I'd think this should only count unique macs. Apologies if this is NOT the desired output for this module.