In discovery, ignore 'notPresent' interfaces
During discovery, ignore all interfaces of type 'ethernetCsmacd'
which are described as 'notPresent'.

The CiscoSB class devices will pre-define phantom interface place-
holders in the MIB for any physical interface which /might/ be in
the stack.  In a large stack this may lead to hundreds of such
phantom interfaces, which just take up vertical space in the GUI
and rows in the DB.

By checking for type == ethernetCsmacd, we allow the Port-Channel
and Tunnel interfaces, which also get described as notPresent, to
appear.