netdisco-docker triage
point netdisco docker users to the docker issue tracker as a first step, so we can spent less time on triaging issues.

netdisco docker has some quirks not present in regular netdisco, if it turns out to be non container related we can always move the issue back to this tracker.

btw, seems we need to change our issue templates, but only project admins can do so:
"You are using an old version of issue templates. Please update to the new issue template workflow. Learn more."
-> https://help.github.com/articles/about-issue-and-pull-request-templates